Don't open realm when deleting it

+++ Realm/Realm/Realm.cs
@@ -17,6 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -161,17 +162,15 @@ namespace Realms
         /// <param name="configuration">A <see cref="RealmConfigurationBase"/> which supplies the realm path.</param>
         public static void DeleteRealm(RealmConfigurationBase configuration)
         {
-            using (var realm = GetInstance(configuration))
+            var fullpath = configuration.DatabasePath;
+            if (RealmState.TryGetState(fullpath, out var state) &&
+                state.RealmCount > 0)
             {
-                if (realm._state.RealmCount > 1)
-                {
-                    throw new RealmPermissionDeniedException("Unable to delete Realm because it is still open.");
-                }
+                throw new RealmPermissionDeniedException("Unable to delete Realm because it is still open.");
             }
 
-            var fullpath = configuration.DatabasePath;
             File.Delete(fullpath);
-            File.Delete(fullpath + ".log_a");  // eg: name at end of path is EnterTheMagic.realm.log_a   
+            File.Delete(fullpath + ".log_a");  // eg: name at end of path is EnterTheMagic.realm.log_a
             File.Delete(fullpath + ".log_b");
             File.Delete(fullpath + ".log");
             File.Delete(fullpath + ".lock");
@@ -223,17 +222,20 @@ namespace Realms
                 state = GCHandle.FromIntPtr(statePtr).Target as RealmState;
             }
 
+            Config = config;
+
             if (state == null)
             {
-                state = new RealmState();
+                state = new RealmState(this);
                 sharedRealmHandle.SetManagedStateHandle(GCHandle.ToIntPtr(state.GCHandle));
             }
-
-            state.AddRealm(this);
+            else
+            {
+                state.AddRealm(this);
+            }
             _state = state;
 
             SharedRealmHandle = sharedRealmHandle;
-            Config = config;
 
             Metadata = schema.ToDictionary(t => t.Name, CreateRealmObjectMetadata);
             Schema = schema;
@@ -618,8 +620,8 @@ namespace Realms
         /// </summary>
         /// <example>
         /// <code>
-        /// using (var trans = realm.BeginWrite()) 
-        /// { 
+        /// using (var trans = realm.BeginWrite())
+        /// {
         ///     realm.Add(new Dog
         ///     {
         ///         Name = "Rex"
@@ -637,18 +639,18 @@ namespace Realms
         }
 
         /// <summary>
-        /// Execute an action inside a temporary <see cref="Transaction"/>. If no exception is thrown, the <see cref="Transaction"/> 
+        /// Execute an action inside a temporary <see cref="Transaction"/>. If no exception is thrown, the <see cref="Transaction"/>
         /// will be committed.
         /// </summary>
         /// <remarks>
-        /// Creates its own temporary <see cref="Transaction"/> and commits it after running the lambda passed to <c>action</c>. 
-        /// Be careful of wrapping multiple single property updates in multiple <see cref="Write"/> calls. 
+        /// Creates its own temporary <see cref="Transaction"/> and commits it after running the lambda passed to <c>action</c>.
+        /// Be careful of wrapping multiple single property updates in multiple <see cref="Write"/> calls.
         /// It is more efficient to update several properties or even create multiple objects in a single <see cref="Write"/>,
         /// unless you need to guarantee finer-grained updates.
         /// </remarks>
         /// <example>
         /// <code>
-        /// realm.Write(() => 
+        /// realm.Write(() =>
         /// {
         ///     realm.Add(new Dog
         ///     {
@@ -678,14 +680,14 @@ namespace Realms
         /// </summary>
         /// <remarks>
         /// Opens a new instance of this Realm on a worker thread and executes <c>action</c> inside a write <see cref="Transaction"/>.
-        /// <see cref="Realm"/>s and <see cref="RealmObject"/>s are thread-affine, so capturing any such objects in 
+        /// <see cref="Realm"/>s and <see cref="RealmObject"/>s are thread-affine, so capturing any such objects in
         /// the <c>action</c> delegate will lead to errors if they're used on the worker thread. Note that it checks the
         /// <see cref="SynchronizationContext"/> to determine if <c>Current</c> is null, as a test to see if you are on the UI thread
         /// and will otherwise just call Write without starting a new thread. So if you know you are invoking from a worker thread, just call Write instead.
         /// </remarks>
         /// <example>
         /// <code>
-        /// await realm.WriteAsync(tempRealm =&gt; 
+        /// await realm.WriteAsync(tempRealm =&gt;
         /// {
         ///     var pongo = tempRealm.All&lt;Dog&gt;().Single(d =&gt; d.Name == "Pongo");
         ///     var missis = tempRealm.All&lt;Dog&gt;().Single(d =&gt; d.Name == "Missis");
@@ -1089,6 +1091,8 @@ namespace Realms
         {
             #region static
 
+            private static readonly ConcurrentDictionary<string, WeakReference<RealmState>> _states = new ConcurrentDictionary<string, WeakReference<RealmState>>();
+
             [NativeCallback(typeof(NativeCommon.NotifyRealmCallback))]
             public static void NotifyRealmChanged(IntPtr stateHandle)
             {
@@ -1096,19 +1100,31 @@ namespace Realms
                 ((RealmState)gch.Target).NotifyChanged(EventArgs.Empty);
             }
 
+            public static bool TryGetState(string path, out RealmState state)
+            {
+                state = null;
+                return _states.TryGetValue(path, out var weakState) &&
+                       weakState.TryGetTarget(out state);
+            }
+
             #endregion
 
-            private readonly List<WeakReference<Realm>> weakRealms = new List<WeakReference<Realm>>();
+            private readonly List<WeakReference<Realm>> _weakRealms = new List<WeakReference<Realm>>();
+            private readonly string _databasePath;
 
             public readonly GCHandle GCHandle;
             public readonly Queue<Action> AfterTransactionQueue = new Queue<Action>();
 
             public int RealmCount => GetLiveRealms().Count();
 
-            public RealmState()
+            public RealmState(Realm realm)
             {
                 // this is freed in a native callback when the CSharpBindingContext is destroyed
                 GCHandle = GCHandle.Alloc(this);
+                _databasePath = realm.Config.DatabasePath;
+                _states.TryAdd(_databasePath, new WeakReference<RealmState>(this));
+
+                AddRealm(realm);
             }
 
             private void NotifyChanged(EventArgs e)
@@ -1123,25 +1139,28 @@ namespace Realms
             {
                 // We only want to have each realm once. That should be the case as AddRealm
                 // is only called in the Realm ctor, but let's check just in case.
-                Debug.Assert(!weakRealms.Any(r =>
+                Debug.Assert(!_weakRealms.Any(r =>
                 {
                     return r.TryGetTarget(out var other) && ReferenceEquals(realm, other);
                 }), "Trying to add a duplicate Realm to the RealmState.");
 
-                weakRealms.Add(new WeakReference<Realm>(realm));
+                Debug.Assert(realm.Config.DatabasePath == _databasePath, "Realm and state have different paths.");
+
+                _weakRealms.Add(new WeakReference<Realm>(realm));
             }
 
             public void RemoveRealm(Realm realm)
             {
-                var weakRealm = weakRealms.SingleOrDefault(r =>
+                var weakRealm = _weakRealms.SingleOrDefault(r =>
                 {
                     return r.TryGetTarget(out var other) && ReferenceEquals(realm, other);
                 });
-                weakRealms.Remove(weakRealm);
+                _weakRealms.Remove(weakRealm);
 
-                if (!weakRealms.Any())
+                if (!_weakRealms.Any())
                 {
                     realm.SharedRealmHandle.CloseRealm();
+                    _states.TryRemove(_databasePath, out var _);
                 }
             }
 
@@ -1149,7 +1168,7 @@ namespace Realms
             {
                 var realms = new List<Realm>();
 
-                weakRealms.RemoveAll(r =>
+                _weakRealms.RemoveAll(r =>
                 {
                     if (r.TryGetTarget(out var realm))
                     {
